dags/requirements.txt
telebot
PyDrive
pydub
SpeechRecognition
torchaudio
gTTS
openai-whisper
accelerate
langdetect
librosa
torch
sentencepiece
postgres
apache-airflow
sacremoses
huggingface_hub
azure-servicebus
psycopg2-binary
pytest-shutil
transformers